requirements.txt
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.38.0
idna==3.4
imageio==2.23.0
kiwisolver==1.4.4
matplotlib==3.6.2
methodtools==0.4.5
networkx==2.8.8
numpy==1.24.1
packaging==22.0
Pillow==9.4.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.28.1
scikit-image==0.19.3
scimage==0.1.7
scipy==1.10.0
six==1.16.0
tifffile==2022.10.10
tinycss2==1.2.1
urllib3==1.26.13
Wand==0.6.11
webencodings==0.5.1
wirerope==0.4.6
